pyproject.toml
[tool.poetry]
name = "skycruiseapi"
version = "0.1.0"
description = ""
authors = ["Aymen <bestaouiaymene@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework-simplejwt = "^5.3.1"
django = "^5.0.4"
djangorestframework = "^3.15.1"
pyyaml = "^6.0.1"
django-split-settings = "^1.3.1"
pydantic = "^1.10.9"
pynacl = "^1.5.0"
django-filter = "^24.2"
pillow = "^10.3.0"
dj-database-url = "^2.1.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
colorlog = "^6.8.2"
pytest = "^7.4.0"
pytest-django = "^4.5.2"
pytest-xdist = "^3.3.1"
model-bakery = "^1.12.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "skycruise.project.settings"
django_find_project = false
pythonpath = [ "." ]
python_files = "test_*.py"
testpaths = ["skycruise"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
